backend/requirements.txt
gunicorn
eventlet
requests
flask
pymongo
libgen_api
flask_cors
pymupdf
ebooklib
tqdm
celery
python-dotenv
redis
Flask-SocketIO
openai
python-pptx